test(unit-tests): refactor dtos used in tests

+++ restaurant-service/src/main/java/com/fooddelivery/restaurant/dto/MenuItemRequestDto.java
@@ -3,12 +3,10 @@ package com.fooddelivery.restaurant.dto;
 import jakarta.validation.constraints.NotBlank;
 import jakarta.validation.constraints.NotNull;
 import jakarta.validation.constraints.Positive;
-import java.math.BigDecimal;
 
 public record MenuItemRequestDto(
     @NotBlank String name,
     String description,
-    @NotNull @Positive BigDecimal price,
+    @NotNull @Positive Integer price,
     String category,
-    String imageUrl
-) {}
+    String imageUrl) {}
